check that coordinates are of type float64 in comparison (#36)
* check that coordinates are of type float64 in comparison

* fix tests

* update tess to have float64 cols

+++ catalog_functions/vipers_transformer.py
@@ -37,7 +37,7 @@ class VIPERSTransformer(BaseTransformer):
         for f in self.FLOAT_FEATURES:
             fields.append(pa.field(f, pa.float32()))
         for f in self.ASTROMETRY_FEATURES:
-            fields.append(pa.field(f, pa.float32()))
+            fields.append(pa.field(f, pa.float64()))
 
         # Object ID
         fields.append(pa.field("object_id", pa.string()))
